[ft] Use TS & Add package.json
